Day 14: Scope
The absolute difference between two integers,  a and b, is written as |a-b| . The maximum absolute difference between two integers in a set of positive integers, elements, is the largest absolute difference between any two integers in elements.

The Difference class is started for you in the editor. It has a private integer array (elements) for storing N non-negative integers, and a public integer (maximumDifference) for storing the maximum absolute difference.

Task
Complete the Difference class by writing the following:
A class constructor that takes an array of integers as a parameter and saves it to the  elements instance variable.
A computeDifference method that finds the maximum absolute difference between any  2 numbers in N and stores it in the  maximumDifference instance variable.
Input Format:
You are not responsible for reading any input from stdin. The locked Solution class in your editor reads in  2 lines of input; the first line contains N, and the second line describes the  elements array.