Day 25: Running Time and Complexity
Task
A prime is a natural number greater than  1 that has no positive divisors other than  1 and itself. Given a number, n, determine and print whether it's  Prime or Not prime.

Note: If possible, try to come up with a  0(Square root n) primality algorithm, or see what sort of optimizations you come up with for an  0(n) algorithm. Be sure to check out the Editorial after submitting your code!